Changed _energy to add user-defined form
It would be nice to have symbolic differentiation for this, since it can be tedious to have to calculate every time, then update the code/provide the written form.

Think the best way is to use the self thing (class? Need to look at this next) to hold the parameter for the form of the energy so that this can be found in _energy and the derivatives in partial_deriv.

Failing this, it can just ask the user for the form of the derivative.

Also, the parameters should be extracted from self too, since you can't pass these in. Need to work out how to do this.