Seems to work for the Yeoh model
Need to check that results are consistent with Yeoh model for different parameter values/ranges. Then also need to check with other models.

Not updated the _energy function yet. Not sure when or how it is used.

Added default values for the Yeoh model as comments at the end of the lines for user inputs. This makes debugging for this model a bit easier.

Added an energy_form attribute to the material. This holds the user-defined string to define the form of the strain energy density function (SEDF).

Added a param_names attribute to the material when the "Ross" material is used. Same scenario for the Yeoh model form here for debugging. Only limitations on the form of the function are that it must rely on I1, I2, I3, and/or I4, that the derivatives must be able to be evaluated by sympy, and that the user must define at least an initial guess for the material constants.

Updated partial_deriv function to only take arguments of self and extra_args. All material parameters are passed in via extra_args (since we don't know their names or order). These are then sorted using the param_names attributes of self.

Not sure if I need to import sympy here. Something to discuss with Ankush. Might be why I think it is so slow (although maybe I'm impatient). The invariants up to I4 are made symbols. The SEDF is evaluated to get the value using the parameters passed in. The derivatives are taken and stored in dSEDFdI.  This is then converted to a string to allow the expressions of In to be replaced with the correct format of self.In. This is done for I1 to I4. The value of the derivative is evaluated to be a floating point value, before replacing the zeros with Nones.

Also need to pass the self.param_names in to the partial_deriv function, since these are required. This causes no issues with older code, since they simply don't use this attribute.