Noticed issues when using this for the LS model
Because of the exponential function in the LS model, needed to handle this. Now converts "exp()" to "sp.exp()" in the form of the energy before evaluating this.

Also need to replace the "exp()" in the resulting symbolic derivatives with "np.exp()".

Previously had that if the derivative was zero, this value should be set to None. However, for the LS model, when I1=3, the derivative is actually zero. Had to implement a check in a for loop. Not managed to write this in list comprehension yet.

Also need to discuss why the LS model form looks different in the paper and in the code. Will test with another material model in the meantime.