Added tested error handling
Mainly pertaining to user inputs. Every time the user is asked for an input, need to be ready for mistakes (within reason).

Also allows the user to avoid giving bounds for parameters, since they are not necessarily used.

Added a success flag. This continues to execute code until satisfied. Also gives warnings to user if the bounds are not set.