"Offline" symbolic derivative
Moved the evaluation of the symbolic derivative outside of the partial derivative function since this is often called multiple times. Instead, I just pass in the form of the partial derivative as a string.

Started commenting some lines. This will need completed at some point.

Updated the _energy function although this has not been tested yet.

Calculate the derivatives when constructing the class, so only done once. To achieve this, needed to specify parameters as symbols since numerical values not yet known.

New syntax for evaluating the partial derivative since this only needs evaluating with the new values.