First steps in making a sparse regression script
Based on sequential thesholding of a least squares fit, like SINDy.

Added a function that generates initial guess, lower, and upper bounds for the parameters. Saves time instead of doing it manually. Might need updated so that this can be entered by the user or provided via a string instead.

Defined some variables that satisfy the minima requirements at I1=3 and I4 = 1. These are replaced in the strings but allows functions to be formed as f(x,y).

Added some comments to outline sort of functions that should be included in our library. Decided to have linear and nonlinear forms up to ^3.
Also decided to add all of the material models included in pymecht since a combination of these could be the best option.

How this works is that it forms a string with all different linear and nonlinear terms, as well as initialising guesses, lower, and upper bounds for these "linear coefficients" (coefficients which multiply terms in a linear manner, denoted ltheta. Idea is to use these to check for sparsity).

Remove the zero order terms. Change the first order terms to just be the variable. Replace the variables x and y with their respective strings.

Currently input own terms as a string required to be copy and pasted. Might change this at some point. Skips the first 17. 21. 20, and 12 characters of the initial guess, lower bound, upper bound, and energy, respectively. This is just due to additional (constant) terms at the beginning. This should be better handled at some point.

Added a sparse_fit model option with all models, plus arb.

Because we have multiple models, need to loop over all models adding the fibre direction for each.