First success with loading data
Just added constant I1 and I4 data at the moment. Ultimately, all should be loaded and potentially give the user the option to configure the choice of data.

Needed to add the pandas package to handle loading from Excel.

Instead of loading the input, output, and protocols data from a npz file, these are read from the Excel spreadsheet, formatted (for each protocol, need to append stretch and stress data in each dimension). Kept the commented lines from the old code since this reminds me that the inputs are stretches and the outputs are stresses.

Added the timer back in.