Added some plans and linear params list
Added a linear_params list. This will be used to search for parameters that should be set to 0 if they are in the linear params list.

All but the first (ltheta_00*1*1) of the power series list are added to the L_params list.

Added the linear parameters from the established MatModels. Will need to add setting bounds here as well.

Added a fixed parameters list that can be updated as we test for sparsity.

Worked out where the loop should start.

Some preliminary planning for the sparsity check.