Altered derivatives in arb MatModel
Now manipulate the expression to avoid terms like: D(mu*(I1-3)^n) = mu*n*(I1-3)**n/(I1-3). This is the same as mu*n*(I1-3)**(n-1) only this suffers from the issue where I1=3, so that inf is obtained. For the new expression, the inverse terms are incorporated in the powers and are only ever an issue if the argument is negative (see issues with 0.25:1 biaxial data).