Working sequential thresholding
Using a tolerance of 0.1 at the moment. This means that terms with linear parameters below 10% of the max are removed from the fit.

Added a condition to avoid a term in the energy where X**0*Y**0 without slicing string.

Added a lower bound on the nlthetap parameters of 1.0. This avoids scenarios where the derivative of X**0.5 becomes 0.5*X**(-0.5). This is a problem when X=0.

Avoided slicing strings again since no term at the beginning.

Added loop for different models. This sets the fibre directions of all models to be [1,0,0].

Also added a loop for sequential thresholding. This has a list of fixed_params. The while loop exits when the number of fixed parameters after removal is the same as before.