Updated arb model to take input variables
Needed to change how this is called in script. Now need to instantiate this class in a variable and add this variable as an argument to the MatModel class.

Needed more error handling to check if the string is empty before asking the user for their value. If a value isn't provided, this asks the user for a value.

This happens for the initial guess, the lower bound, the upper bound, and the form of the SEDF.