Committing some working changes
These were changes that were made as I was using the script. Not often major changes, right enough.

Noticed that the Z "subvariable" was the same as Y.

Changed the way that I generate a library. However, this has only been commented to allow this to be added at a later date, should it be desired.

Added the capability to pass a string into the arb class with some syntax.

Added a SEDF string list to check that all entries are unique. Not often used but kept there in case useful in the future.

Made a new library function.