Added capabilities to profile time and memory
Using cProfile to process the time that functions take to run.

Also need some other libraries so that we can process the data and trace the memory usage. Added a function to help with processing the data.

Added all code within a main function so that the memory profiler can profile it.