All code working in parallel
Not able to merge/cherry-pick any of this into the other branch. Diff will most likely offer limited benefits. Will probably be best just to compare/copy and paste.

Added packages for running in parallel. Also added a function to check if code is running in parallel.

Needed to rearrange the code a bit. Brought some functions further up. Added input and sample to the residual function. Added sample to the ForwardDifference function. Same for CentralDifference. Added dsampledtheta to inputs for the Dresidual function.

Rearranged code to ask for the LOGNORM flag first. Added syntax to enable running in parallel.

Added some user inputs to the top of the ___main___ loop.