Added powdenest
Added this function as it seems to speed up the evaluation of the derivative through reducing the length of the expression. Does not seem to have any negative effects apart from meaning that simpler forms without many exponential terms are slower to fit. These are generally quick though, so not a real issue.