Changes to allow flexibility with derivatives
Need to account for the case that dSEDFdI1 is a function of each value of I_{4,i}. That is, that dSEDFdI1 will be a list len(self.M) long. As such, the contribution is also a summation of these. Similarly, need to account for the case that dSEDFdI4 is just a single, scalar value (I1*I4, for example). This still needs summed but we do not need to index over the value of dSEDFdI4, since they are all identical.