Recent (marginal) improvements
Changed the default upper and lower bounds to be +/-1000000.

Added a user input for using bounds.

Added a library of previous best fit values that can be used as an initial guess. Sometimes this really speeds up the fit if they are a good guess.

Added syntax to use the previous guess if the parameter appears in the library. If it does not, this is initialised with the default values. This happens for both the linear and nonlinear parameters.

Added code to check if bounds are used.