Added problem parameter sets
Here, since the derivative wrt I1 is a function of I4 and vice versa, the problem arises. This is not an issue for dPsidI1 if only fibre direction is chosen. However, dPsidI4 is always an issue with these parameters as it is only a function of I1, so is not iterable.