Rearranging the structure of repo
Now have the toml file in the top level of the repo with MatModel, RandomParameters, and SampleExperiment modules in their own folders with their own __init__.py files. The LinearFEM.py file is currently in the SampleExperiment folder but will probably be removed at a later date.

Needed to add the modules to the toml file using the packages option.

Added an Examples folder to hold all examples.